Remove redundant template
