+++ src/query/parallel/px_worker_manager_global.hpp
@@ -54,7 +54,7 @@ namespace parallel_query
       friend class worker_manager;
 
       /* member variables (ordered by size for alignment) */
-      cubthread::entry_workpool *m_worker_pool;
+      cubthread::worker_pool *m_worker_pool;
       std::once_flag m_init_flag;
       std::atomic<int> m_available;
       int m_capacity;
