+++ src/transaction/log_impl.h
@@ -954,7 +954,7 @@ extern bool cdc_Logging;
 
 #if defined (SERVER_MODE)
 // *INDENT-OFF*
-extern cubthread::entry_workpool *g_backup_read_worker_pool;
+extern cubthread::worker_pool *g_backup_read_worker_pool;
 // *INDENT-ON*
 #endif
